Country Search app complete for the most part
The app is complete when you search for a country it shows you the information about the country but it also shows you an image however improvements can be made. Sometimes the image doesn't properly fit on the screen and there is also no error handling done for when the user searches for an invalid country. And there are no animations or moving parts.